requirements.txt
flake8==2.5.4
mccabe==0.4.0
mock==1.3.0
numpy==1.10.4
pbr==1.8.1
pep8==1.7.0
py==1.4.31
pyepics==3.2.5rc3
pyflakes==1.0.0
pyrobot==0.1.0
pytest==2.9.0
pyzmq==15.2.0
six==1.10.0
